+++ src/main/java/io/bootique/linkmove/LinkMoveRuntimeBuildCallback.java
@@ -1,7 +1,0 @@
-package io.bootique.linkmove;
-
-import com.nhl.link.move.runtime.LmRuntimeBuilder;
-
-public interface LinkMoveRuntimeBuildCallback {
-	void build(LmRuntimeBuilder builder);
-}